// backend/src/app/common/enums.ts
export enum ROLES {

  SA = 'SA',
  ADMIN = 'ADMIN',
  MD = 'MD',
  MODERATOR = 'MODERATOR',
  // CRM = 'Customer Manager',

  // PN = 'Partner',
  SH = 'SH',
  STAKEHOLDER = 'STAKEHOLDER',
  LD = 'LD',
  LEAD = 'LEAD',
  CONTACT = 'CONTACT',

  // TA = 'Talent Acquisition Team',

  HM = 'HM',
  HIRINGMANAGER = 'HIRINGMANAGER',

  HR = 'Human Resources',
  RECRUITER = 'RECRUITER',
  RC = 'RC',
  // HH = 'Head Hunter',

  // SE = "Skills Evaluator",
  TALENT = 'TALENT',
  // TJA = 'Talent / Job Applicant',
  // TJAM = 'Talent / Job Applicant / Manager',
  // TJASME = 'Talent / Job Applicant / Subject Expert',

  VISITOR = 'VISITOR',

  // //
  // CANDIDATE = '',
  // JA = '',
  // CS = '',
}

export enum VerificationType {
  Email = 'Email',
  Phone = 'Phone'
}